refactor: group core base classes
